+++ docs/README.md
@@ -38,7 +38,7 @@
 - [x] input창을 공백으로 만든다. 
 
 # ✅ 요구 사항 
-- [] `play` 메서드를 만들어서 게임을 진행한다. 
+- [x] `play` 메서드를 만들어서 게임을 진행한다. 
   - play 메서드는 `String`으로 결과를 return 한다. (볼을 스트라이크보다 먼저 쓰기)
   - index.js에서 BaseballGame() 생성자 함수 형태를 활용한다. (리드미 참고)
 - [x] 게임을 다시 시작하는 재시작 button 태그는 `game-restart-button` id를 가진다.
@@ -70,7 +70,7 @@
 - [x] 복사본 만들어놓기
 - [x] eslint & prettier (자바스크립트 코드 컨벤션을 지키면서 프로그래밍 한다.)
 - [x] 메소드 길이 -countAnswer
-- [] 파일 분리 (아이패드 참고)
+- [x] 파일 분리 (아이패드 참고)
 - [] 주석 삭제
 - [] 마지막에 요구사항, 기능요구사항 다시한번 체크(완료)
